feat(fat16-open): implemented open function for fat 16

+++ src/macros.h
@@ -2,7 +2,7 @@
 #define MACROS_H
 #include "stdint.h"
 
-#define ISERR(val) ((uintptr_t)val < 0)
+#define ISERR(val) ((int)(uintptr_t)val < 0)
 #define ISERR_I(val) (val < 0)
 #define ERROR(val) (void *)(val)
 #define ERROR_I(val) (uintptr_t)(val)
